html-imports/link-import-rel.html

<!DOCTYPE html>.
<html>
<head>  
<title>HTML Imports Rel</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="author" title="Sangsoo Nam" href="mailto:sangsoo.ted@gmail.com.com">
<link rel="help" href="https://dvcs.w3.org/hg/webcomponents/raw-file/tip/spec/imports/index.html#loading-imports">
<link id="linkImportMissingRel" href="resources/hello.html">
</head>
<body>
<div id="log">Running test...</div>
<script>

test(function() {
  var linkImport = document.getElementById("linkImportMissingRel");
  assert_equals(linkImport.import, null, "When the link link does not represent an import, it should be null.");
}, 
"Check on an import rel attribute. The rel attribute of a link is missing so the link does not represent an import."
);

</script>
</body>
</html>
